+++ src/main/java/mate/academy/spring/controller/CinemaHallController.java
@@ -6,8 +6,8 @@ import mate.academy.spring.model.CinemaHall;
 import mate.academy.spring.model.dto.request.CinemaHallRequestDto;
 import mate.academy.spring.model.dto.response.CinemaHallResponseDto;
 import mate.academy.spring.service.CinemaHallService;
-import mate.academy.spring.service.dto.mapping.DtoRequestMapper;
-import mate.academy.spring.service.dto.mapping.DtoResponseMapper;
+import mate.academy.spring.mapper.DtoRequestMapper;
+import mate.academy.spring.mapper.DtoResponseMapper;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
