+++ src/main/java/mate/academy/spring/dao/impl/ShoppingCartDaoImpl.java
@@ -44,7 +44,7 @@ public class ShoppingCartDaoImpl extends AbstractDao<ShoppingCart> implements Sh
             if (transaction != null) {
                 transaction.rollback();
             }
-            throw new DataProcessingException("Cannot create shopping cart ", e);
+            throw new DataProcessingException("Cannot update shopping cart ", e);
         } finally {
             if (session != null) {
                 session.close();
