+++ src/main/java/mate/academy/spring/exception/AuthenticationException.java
@@ -1,7 +1,6 @@
 package mate.academy.spring.exception;
 
 public class AuthenticationException extends Exception {
-
     public AuthenticationException(String message) {
         super(message);
     }
