+++ src/main/java/mate/academy/spring/mapper/DtoRequestMapper.java
@@ -1,4 +1,4 @@
-package mate.academy.spring.service.dto.mapping;
+package mate.academy.spring.mapper;
 
 public interface DtoRequestMapper<D, C> {
     C fromDto(D dto);
