+++ src/main/java/mate/academy/spring/mapper/impl/request/MovieRequestMapper.java
@@ -1,8 +1,8 @@
-package mate.academy.spring.service.dto.mapping.impl.request;
+package mate.academy.spring.mapper.impl.request;
 
 import mate.academy.spring.model.Movie;
 import mate.academy.spring.model.dto.request.MovieRequestDto;
-import mate.academy.spring.service.dto.mapping.DtoRequestMapper;
+import mate.academy.spring.mapper.DtoRequestMapper;
 import org.springframework.stereotype.Component;
 
 @Component
