+++ src/main/java/mate/academy/spring/mapper/impl/response/MovieSessionResponseMapper.java
@@ -1,9 +1,9 @@
-package mate.academy.spring.service.dto.mapping.impl.response;
+package mate.academy.spring.mapper.impl.response;
 
 import java.time.format.DateTimeFormatter;
 import mate.academy.spring.model.MovieSession;
 import mate.academy.spring.model.dto.response.MovieSessionResponseDto;
-import mate.academy.spring.service.dto.mapping.DtoResponseMapper;
+import mate.academy.spring.mapper.DtoResponseMapper;
 import org.springframework.stereotype.Component;
 
 @Component
