+++ src/main/java/mate/academy/spring/model/CinemaHall.java
@@ -15,9 +15,6 @@ public class CinemaHall {
     private int capacity;
     private String description;
 
-    public CinemaHall() {
-    }
-
     public Long getId() {
         return id;
     }
