+++ src/main/java/mate/academy/spring/model/Movie.java
@@ -15,13 +15,6 @@ public class Movie {
     private String title;
     private String description;
 
-    public Movie() {
-    }
-
-    public Movie(String title) {
-        this.title = title;
-    }
-
     public Long getId() {
         return id;
     }
