+++ src/main/java/mate/academy/spring/model/MovieSession.java
@@ -21,9 +21,6 @@ public class MovieSession {
     private CinemaHall cinemaHall;
     private LocalDateTime showTime;
 
-    public MovieSession() {
-    }
-
     public Long getId() {
         return id;
     }
