+++ src/main/java/mate/academy/spring/model/User.java
@@ -49,4 +49,12 @@ public class User {
     public void setSalt(byte[] salt) {
         this.salt = salt;
     }
+
+    @Override
+    public String toString() {
+        return "User{"
+                + "id=" + id
+                + ", email='" + email + '\''
+                + '}';
+    }
 }
