+++ src/main/java/mate/academy/spring/util/HashUtil.java
@@ -5,19 +5,22 @@ import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 
 public class HashUtil {
-    public static final String HASH_ALGORITHM = "SHA-512";
-    public static final int CYPHER_INTENSITY = 16;
+    private static final String HASH_ALGORITHM = "SHA-512";
+    private static final int CYPHER_INTENSITY = 16;
+
+    private HashUtil() {
+    }
 
     public static String hashPassword(String password, byte[] salt) {
-        StringBuilder hashedPwd = new StringBuilder();
+        StringBuilder hashedPassword = new StringBuilder();
         try {
             MessageDigest messageDigest = MessageDigest.getInstance(HASH_ALGORITHM);
             messageDigest.update(salt);
-            byte[] digest = messageDigest.digest(password.getBytes());
-            for (byte b : digest) {
-                hashedPwd.append(String.format("%02x", b));
+            byte[] symbols = messageDigest.digest(password.getBytes());
+            for (byte symbol : symbols) {
+                hashedPassword.append(String.format("%02x", symbol));
             }
-            return hashedPwd.toString();
+            return hashedPassword.toString();
         } catch (NoSuchAlgorithmException e) {
             throw new RuntimeException("Can`t hash password!", e);
         }
